+++ lib/rubocop/cop/metrics/line_length.rb
@@ -138,7 +138,12 @@ module RuboCop
 
         def ignored_line?(line, line_index)
           matches_ignored_pattern?(line) ||
+            shebang?(line, line_index) ||
             heredocs && line_in_permitted_heredoc?(line_index.succ)
+        end
+
+        def shebang?(line, line_index)
+          line_index.zero? && line.start_with?('#!')
         end
 
         def register_offense(loc, line, line_index)