+++ lib/rubocop/cop/style/guard_clause.rb
@@ -37,9 +37,10 @@ module RuboCop
       #   ok
       class GuardClause < Cop
         include MinBodyLength
+        include StatementModifier
 
-        MSG = 'Use a guard clause instead of wrapping the code inside a ' \
-              'conditional expression.'
+        MSG = 'Use a guard clause (`%<example>s`) instead of wrapping the ' \
+              'code inside a conditional expression.'
 
         def on_def(node)
           body = node.body
@@ -55,9 +56,19 @@ module RuboCop
         alias on_defs on_def
 
         def on_if(node)
-          return if accepted_form?(node) || !contains_guard_clause?(node)
+          return if accepted_form?(node)
 
-          add_offense(node, location: :keyword)
+          guard_clause_in_if = node.if_branch&.guard_clause?
+          guard_clause_in_else = node.else_branch&.guard_clause?
+          guard_clause = guard_clause_in_if || guard_clause_in_else
+          return unless guard_clause
+
+          kw = if guard_clause_in_if
+                 node.loc.keyword.source
+               else
+                 opposite_keyword(node)
+               end
+          register_offense(node, guard_clause.source, kw)
         end
 
         private
@@ -65,7 +76,30 @@ module RuboCop
         def check_ending_if(node)
           return if accepted_form?(node, true) || !min_body_length?(node)
 
-          add_offense(node, location: :keyword)
+          register_offense(node, 'return', opposite_keyword(node))
+        end
+
+        def opposite_keyword(node)
+          node.if? ? 'unless' : 'if'
+        end
+
+        def register_offense(node, scope_exiting_keyword, conditional_keyword)
+          condition, = node.node_parts
+          example = [scope_exiting_keyword,
+                     conditional_keyword,
+                     condition.source].join(' ')
+          if too_long_for_single_line?(node, example)
+            example = "#{conditional_keyword} #{condition.source}; " \
+                      "#{scope_exiting_keyword}; end"
+          end
+          add_offense(node,
+                      location: :keyword,
+                      message: format(MSG, example: example))
+        end
+
+        def too_long_for_single_line?(node, example)
+          max = max_line_length
+          max && node.source_range.column + example.length > max
         end
 
         def accepted_form?(node, ending = false)
@@ -80,11 +114,6 @@ module RuboCop
           else
             !node.else? || node.elsif?
           end
-        end
-
-        def contains_guard_clause?(node)
-          node.if_branch&.guard_clause? ||
-            node.else_branch&.guard_clause?
         end
       end
     end